+++ core/lib/Thelia/Coupon/Type/CouponInterface.php
@@ -221,13 +221,6 @@ interface CouponInterface
     public function drawBackOfficeInputs();
 
     /**
-     * Get all extended inputs name to manage
-     *
-     * @return mixed
-     */
-    public function getExtendedInputs();
-
-    /**
      * @return ObjectCollection list of country IDs for which shipping is free. All if empty
      */
     public function getFreeShippingForCountries();
@@ -236,4 +229,14 @@ interface CouponInterface
      * @return ObjectCollection list of module IDs for which shipping is free. All if empty
      */
     public function getFreeShippingForModules();
+
+    /**
+     * Create the effect array from the list of fields
+     *
+     * @param array $data the input form data (e.g. $form->getData())
+     *
+     * @return array a filedName => fieldValue array
+     */
+    public function getEffects($data);
+
 }