+++ core/lib/Thelia/Coupon/Type/RemoveXAmount.php
@@ -12,6 +12,8 @@
 
 namespace Thelia\Coupon\Type;
 
+use Thelia\Core\Translation\Translator;
+
 /**
  * Allow to remove an amount from the checkout total
  *
@@ -25,9 +27,7 @@ class RemoveXAmount extends CouponAbstract
     protected $serviceId = 'thelia.coupon.type.remove_x_amount';
 
     /**
-     * Get I18n name
-     *
-     * @return string
+     * @inheritdoc
      */
     public function getName()
     {
@@ -37,21 +37,7 @@ class RemoveXAmount extends CouponAbstract
     }
 
     /**
-     * Get I18n amount input name
-     *
-     * @return string
-     */
-    public function getInputName()
-    {
-        return $this->facade
-            ->getTranslator()
-            ->trans('Discount amount', array(), 'coupon');
-    }
-
-    /**
-     * Get I18n tooltip
-     *
-     * @return string
+     * @inheritdoc
      */
     public function getToolTip()
     {
@@ -66,12 +52,43 @@ class RemoveXAmount extends CouponAbstract
         return $toolTip;
     }
 
+    /**
+     * @inheritdoc
+     */
     public function drawBackOfficeInputs()
     {
         return $this->facade->getParser()->render('coupon/type-fragments/remove-x-amount.html', [
-                'label'     => $this->getInputName(),
-                'fieldName' => self::INPUT_AMOUNT_NAME,
+                'fieldName' => $this->makeCouponFieldName(self::AMOUNT_FIELD_NAME),
                 'value'     => $this->amount
             ]);
     }
+
+    /**
+     * @inheritdoc
+     */
+    protected function getFieldList()
+    {
+        return [self::AMOUNT_FIELD_NAME];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function checkCouponFieldValue($fieldName, $fieldValue)
+    {
+        if ($fieldName === self::AMOUNT_FIELD_NAME) {
+
+            if (floatval($fieldValue) < 0) {
+                throw new \InvalidArgumentException(
+                    Translator::getInstance()->trans(
+                        'Value %val for Disount Amount is invalid. Please enter a positive value.',
+                        [ '%val' => $fieldValue]
+                    )
+                );
+            }
+        }
+
+        return $fieldValue;
+    }
+
 }